pyproject.toml
[tool.poetry]
name = "bi-orchestrator"
version = "0.1.0"
description = "Data orchestrator for BI: integrates CRM, Ads and other sources for business analysis."
authors = ["Sebastian Felipe Franco Herrera"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.3.0"
requests = "^2.32.4"
python-dotenv = "^1.1.0"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
